move the copied jsonLinks directly into YLS's existing yamlLinks
